DEV-366 - Clear frozen-scope InspectCode WARNING findings

+++ src/LamuFlix.Web/Models/Helper/PagedListing.cs
@@ -43,7 +43,7 @@ public class PagedListing<T> : IPagedListing
         CurrentPage = page;
 
         // Total Pages
-        TotalPages = (int)(Math.Ceiling((decimal)TotalItems / (decimal)ItemsPerPage));
+        TotalPages = (int)Math.Ceiling((decimal)TotalItems / ItemsPerPage);
 
         // Set PagerStart
         var pagingStart = CurrentPage - 5;
